fix: qHash return type uint → size_t (Qt 6 signature, removes C4267 warning)

+++ src/document/UBDocumentProxy.h
@@ -150,7 +150,7 @@ inline bool operator==(const UBDocumentProxy &proxy1, const UBDocumentProxy &pro
     return proxy1.persistencePath() == proxy2.persistencePath();
 }
 
-inline uint qHash(const UBDocumentProxy &key)
+inline size_t qHash(const UBDocumentProxy &key)
 {
     return qHash(key.persistencePath());
 }
